pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "egwwoa"
version = "0.1.0"
description = "EGW–WOA with Lévy flights and chaotic re-initialization"
authors = [{name="Ercan Erkalkan", email="ercan.erkalkan@marmara.edu.tr"}]
requires-python = ">=3.10"
dependencies = ["numpy","scipy","pandas","pyyaml","matplotlib","tqdm","scikit-learn","tabulate","colorlog"]

[tool.setuptools.packages.find]
where = ["src"]
